feat(ui): v2 opt-in shell — design tokens, router, Beta tab toggle [M]
UI/UX maturity refresh inspired by Wheneri (heavy display titles,
hairline-list aesthetic, single circular-pill close affordance), delivered
as a v2 shell behind an opt-in toggle so v1 stays exactly as-is and users
can flip back any time.

src/App.jsx becomes a thin router reading localStorage.ui_version (default
'v1'). URL escape hatch ?ui=v2 / ?ui=v1 sets the flag and strips itself
from the URL so deep-link params (?task=X from notifications) survive.
data-ui-version is mirrored on the documentElement; v2 also sets
data-ui="v2" so its namespaced --v2-* tokens activate without leaking.

Tokens at src/v2/tokens.css: single accent (#FF6240), muted alerts
(#E8443A overdue, #F2A100 high-pri), pastel-ified energy palette,
off-white #FAFAF7 light background, three named easings (240/180/120ms).
Dark variant keys off existing data-theme="dark".

src/v2/AppV2.jsx is a placeholder welcome page — subsequent PRs (Header,
TaskCard, ModalShell, etc.) replace it. v2 is currently visual-only;
shared hooks, contexts, server endpoints, store, and api.js are reused
as-is.

Settings → new Beta tab hosts the toggle. Reserved for future opt-in
experiments too.

Original src/App.jsx renamed to src/AppV1.jsx with function and export
also renamed; otherwise content untouched.

End state plan: after the 8-PR v2 build-out and a 1-2 week opt-in period,
flip default to 'v2', leave v1 reachable via ?ui=v1 for one release, then
delete AppV1.jsx + components/ and rename v2/components/ → components/.
